+++ Tests/WindowsCloudPC.Tests.ps1
@@ -3,7 +3,8 @@ BeforeDiscovery {
         'Connect-CloudPC',
         'Get-CloudPC',
         'Get-CloudPCProvisioningPolicy',
-        'Get-CloudPCUsage'
+        'Get-CloudPCUsage',
+        'Get-CloudPCUsageBeta'
     )
 }